+++ core-ng-test/src/main/java/core/framework/test/redis/MockRedisStore.java
@@ -3,6 +3,7 @@ package core.framework.test.redis;
 import core.framework.util.Maps;
 
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -49,7 +50,7 @@ class MockRedisStore {
         }
 
         @SuppressWarnings("unchecked")
-        Map<String, String> map() {
+        Map<String, String> hash() {
             assertThat(value).isInstanceOf(Map.class);
             return (Map<String, String>) value;
         }
@@ -60,14 +61,26 @@ class MockRedisStore {
             return (Set<String>) value;
         }
 
-        @SuppressWarnings("unchecked")
-        Map<String, Long> sortedSet() {
-            assertThat(value).isInstanceOf(Map.class);
-            return (HashMap<String, Long>) value;
+        SortedSet sortedSet() {
+            assertThat(value).isInstanceOf(SortedSet.class);
+            return (SortedSet) value;
+        }
+
+        HyperLogLog hyperLogLog() {
+            assertThat(value).isInstanceOf(HyperLogLog.class);
+            return (HyperLogLog) value;
         }
 
         boolean expired(long now) {
             return expirationTime != null && now >= expirationTime;
         }
+    }
+
+    // to distinguish with hash key type
+    static class SortedSet extends HashMap<String, Long> {
+    }
+
+    // use hashset as a naive impl, may be replaced by bloom filter in future
+    static class HyperLogLog extends HashSet<String> {
     }
 }