+++ core-ng-test/src/test/java/core/framework/test/redis/MockRedisHyperLogLogTest.java
@@ -10,34 +10,37 @@ import static org.assertj.core.api.Assertions.assertThat;
  * @author tempo
  */
 public class MockRedisHyperLogLogTest {
-    private RedisHyperLogLog logLog;
+    private RedisHyperLogLog log;
 
     @BeforeEach
-    void setup() {
-        logLog = new MockRedis().hyperLogLog();
+    void createMockRedisHyperLogLog() {
+        log = new MockRedis().hyperLogLog();
     }
 
     @Test
     void add() {
-        assertThat(logLog.add("loglog", "value1")).isTrue();
-        assertThat(logLog.add("loglog", "value1")).isFalse();
+        assertThat(log.add("log1", "value1")).isTrue();
+        assertThat(log.add("log1", "value1")).isFalse();
 
-        assertThat(logLog.add("loglog", "value1", "value2")).isTrue();
-        assertThat(logLog.add("loglog", "value1", "value2")).isFalse();
+        assertThat(log.add("log1", "value1", "value2")).isTrue();
+        assertThat(log.add("log1", "value1", "value2")).isFalse();
     }
 
     @Test
     void count() {
-        assertThat(logLog.count("key1")).isEqualTo(0);
+        assertThat(log.count("key1")).isEqualTo(0);
 
-        logLog.add("key1", "value1");
-        assertThat(logLog.count("key1")).isEqualTo(1);
+        log.add("key1", "value1");
+        assertThat(log.count("key1")).isEqualTo(1);
 
-        logLog.add("key1", "value1", "value2");
-        assertThat(logLog.count("key1")).isEqualTo(2);
+        log.add("key1", "value1", "value2");
+        assertThat(log.count("key1")).isEqualTo(2);
+        assertThat(log.count("key1", "key2")).isEqualTo(2);
 
-        assertThat(logLog.count("key1", "key2")).isEqualTo(2);
-        logLog.add("key2", "value1");
-        assertThat(logLog.count("key1", "key2")).isEqualTo(3);
+        log.add("key2", "value1");
+        assertThat(log.count("key1", "key2")).isEqualTo(2);
+
+        log.add("key2", "value3");
+        assertThat(log.count("key1", "key2")).isEqualTo(3);
     }
 }