+++ core-ng/src/main/java/core/framework/redis/Redis.java
@@ -34,13 +34,13 @@ public interface Redis {
 
     void multiSet(Map<String, String> values);
 
+    void forEach(String pattern, Consumer<String> consumer);
+
     RedisHash hash();
 
     RedisList list();
 
     RedisSortedSet sortedSet();
-
-    void forEach(String pattern, Consumer<String> consumer);
 
     RedisAdmin admin();